requirements.txt
streamlit==1.29.0
groq==0.30.0
python-dotenv==1.0.0 